% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{format.mortaar_life_table}
\alias{format.mortaar_life_table}
\title{Encode a mortaar_life_table in a Common Format}
\usage{
\method{format}{mortaar_life_table}(life_table, class_of_deceased = NULL, ...)
}
\arguments{
\item{class_of_deceased}{optional specify the class of deceased (male, female, undetermined, ...)}

\item{life_table_list}{a mortaar_life_table}
}
\value{
A string representation of the mortaar_life_table
}
\description{
Format an mortaar_life_table for pretty printing.
}
\examples{

test <- data.frame(male = round(runif(70)*100, 0), female = round(runif(70)*100, 0))
format.mortaar_life_table(life.table(test)$male)

}

