% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{format.mortaar_life_table_list}
\alias{format.mortaar_life_table_list}
\title{Encode a mortaar_life_table_list in a Common Format}
\usage{
\method{format}{mortaar_life_table_list}(x, ...)
}
\arguments{
\item{x}{a mortaar_life_table_list}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A string representation of the mortaar_life_table_list. Each format.mortaar_life_table
is formated by its own method and concatenated. The names of the elements are used to specify
the name in the returned header of the printout
}
\description{
Format an mortaar_life_table_list for pretty printing.
}
\examples{

#test <- data.frame(male = round(runif(70)*100, 0), female = round(runif(70)*100, 0))
#format(life.table(test))

}

