% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{function1}
\alias{function1}
\title{Resolves a field (RangeOfAgeFieldName) with the format "1-12"}
\usage{
function1(x, RangeOfAgeFieldName)
}
\arguments{
\item{x}{a dataframe}

\item{RangeOfAgeFieldName}{Column name of the age class (as character)}
}
\value{
dataframe with additional columns "from" and "to"
}
\description{
Resolves a field (RangeOfAgeFieldName) with the format "1-12" to two fields with the numeric values 1 and 12
}
\examples{

#test=read.table("data-raw/siedl.txt",header=T,sep="\\t")
#test1=function1(test,"Jahrefeld")

}

