% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{function3}
\alias{function3}
\title{Creates the input for the function life.table}
\usage{
function3(x, CountOfDeceasedFieldName = "NA", BeginOfAgeFieldName,
  EndOfAgeFieldName, GroupName = "NA", methode = "NA")
}
\arguments{
\item{x}{a dataframe}

\item{CountOfDeceasedFieldName}{Column name of the count of deceased (as character)}

\item{BeginOfAgeFieldName}{Column name of the from field (as character)}

\item{EndOfAgeFieldName}{Column name of the to field (as character)}

\item{GroupName}{Column name of the grouping field (as character)}

\item{methode}{name of the age class determination (as character) Options: "Standard" (default) (1,4,5,5,...), "Equal5" (5,5,...)}
}
\value{
list as input parameter for the function life.table
}
\description{
Creates the input for the function life.table
}
\examples{

#test2=function3(x=test1,CountOfDeceasedFieldName="NA",BeginOfAgeFieldName="from",EndOfAgeFieldName="to",GroupName="Geschlecht", methode="Standard")

}

