% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gallery_graves}
\alias{gallery_graves}
\title{gallery_graves: Life table example}
\format{A data frame with 128 rows and 4 variables.
\itemize{
  \item \bold{indnr:} ID of the individual represented by
                      a skull.
  \item \bold{sex:} sex determination, w -> female, m -> male.
  \item \bold{age:} age class.
  \item \bold{site:} collective grave.
}}
\description{
A dataset from \emph{Czarnetzki 1966}.
}
\details{
Skulls from four different collective graves. This dataset is
not well prepared and needs some adjustment.
}
\references{
\insertRef{czarnetzki_menschlichen_1966}{mortAAR}
}
\seealso{
Other training: \code{\link{magdalenenberg}},
  \code{\link{muensingen}}
}
\concept{training}
