% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{\%+0\%}
\alias{\%+0\%}
\title{Uses the addition function "+" to combine two vectors setting NA values of the first vector to zero.}
\usage{
x \%+0\% y
}
\arguments{
\item{x}{a vector.}

\item{y}{vector to combine with.}
}
\value{
dataframe with additional columns "from" and "to".
}
\description{
Uses the addition function "+" to combine two vectors setting NA values of the first vector to zero.
}
\examples{

c(1, 2, NA, 4) \%+0\% c(5, 6, 7,  8)

}
