% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{is.mortaar_life_table}
\alias{is.mortaar_life_table}
\title{Checks if a variable is of class mortaar_life_table}
\usage{
is.mortaar_life_table(x, ...)
}
\arguments{
\item{x}{a variable}

\item{...}{further arguments passed to or from other methods.}
}
\value{
true if x is a mortaar_life_table, false otherwise
}
\description{
Checks if a variable is of class mortaar_life_table
}
\examples{

#test <- data.frame(male = round(runif(70)*100, 0), female = round(runif(70)*100, 0))
#is.mortaar_life_table(life.table(test)$male)

}

