% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{is.mortaar_life_table_list}
\alias{is.mortaar_life_table_list}
\title{Checks if a variable is of class mortaar_life_table_list}
\usage{
is.mortaar_life_table_list(x, ...)
}
\arguments{
\item{x}{a variable}

\item{...}{further arguments passed to or from other methods.}
}
\value{
true if x is a mortaar_life_table_list, false otherwise
}
\description{
Checks if a variable is of class mortaar_life_table_list
}
\examples{

test <- data.frame(male = round(runif(70)*100, 0), female = round(runif(70)*100, 0))
is.mortaar_life_table_list(life.table(test))

}

