% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{life.table}
\alias{life.table}
\title{Creates a life table}
\usage{
life.table(anzahl)
}
\arguments{
\item{anzahl}{a vector containing number of invidiuals per life year, index references life year[index-1 – index)}
}
\value{
a data frame of class 'mortaar' with the life table variables
}
\description{
simple lifetable using Keyfitz and Flieger separation factors and
exponential tail of death distribution (to close out life table)
partly taken from https://web.stanford.edu/group/heeh/cgi-bin/web/node/75
}

