% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetable_indices.R
\name{lt.indices}
\alias{lt.indices}
\title{Generates indices from a mortAAR life table for use in other functions}
\usage{
lt.indices(life_table)
}
\arguments{
\item{life_table}{an object of class mortaar_life_table.}
}
\value{
A list with the following indices and vectors:
\itemize{
  \item \bold{child_i}:   ratio of individuals aged 5--9 to those aged
  10--14 according to Bocquet-Appel and Masset.
  \item \bold{juvenile_i}:   ratio of individuals aged 5--14 to adults
  according to Bocquet-Appel and Masset.
  \item \bold{P(5-19)}:  ratio of individuals aged 5--19 to those aged
  5 or above according to Bocquet-Appel 2002.
  \item \bold{D0_14_D}:   proportion of individuals aged 0--14
  according to McFadden and Oxenham 2017 if infants are represented
  well.
  \item \bold{mortality}:   list of mortalities 5q0, 5q10, 5q15, 45q15.
  \item \bold{repro_indx}: vector of survival rates of individuals
  aged 15--45.
  \item \bold{e0}:   life expectancy at age 0.
}
}
\description{
This function bundles a number of indices and vectors from a mortAAR
life table which are needed by other functions in mortAAR.
}
