% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/further_analytical_functions.R
\name{lt.reproduction}
\alias{lt.reproduction}
\title{Calculates reproduction indices from a mortAAR life table}
\usage{
lt.reproduction(life_table, fertility_rate = "BA_log", gen_len = 20)
}
\arguments{
\item{life_table}{an object of class mortaar_life_table.}

\item{fertility_rate}{string or numeric. Either fertility rate according
to \emph{McFadden & Oxenham 2018} if infants are represented well or
fertility rate according to data by \emph{McFadden & Oxenham 2018} for
P(5-19) index after \emph{Bocquet-Appel 2002}. Options: 'McO'
(McFadden/Oxenham), 'BA_linear' (linear fit), 'BA_power' (power fit)
or 'BA_log' (logistic fit). Default: BA_log'. Additionally, the
user can specify an arbitrary number in lieu of the fertility rate.}

\item{gen_len}{numeric. Length of generation for determining
the rate of doubling the population. Default: 20.}
}
\value{
A data.frame with basic reproduction indices:

\itemize{
  \item \bold{DR}:  Dependency ratio.

                   \eqn{DR = (sum(D0--14) + sum(D60+)) / sum(D15--59) }

  \item \bold{TFR}:  Total fertility rate.
  \item \bold{GRR}:  Gross reproduction rate.

                   \eqn{GRR = TFR * 0.488}

  \item \bold{NRR}: Net reproduction rate.

                   \eqn{NRR = sum(GRR * age specific fertility * age specific survival / 10000)}

  \item \bold{r}:  Intrinsic growth rate in percent per year.

                   \eqn{r = 100 * log(NRR) / generation length}

  \item \bold{Dt}: Doubling time in years.

                   \eqn{Dt = 100 * 0.6931 / r}
 }
}
\description{
For population studies it is of vital importance to estimate growth or
decline of a population. For archaeological datasets this is rarely
attempted, probably because the data quality seems to scanty.
Nevertheless, the calculation of such measures seems worth the try,
at least it should give an impression if the resulting values
are unrealistic high or low.
}
\details{
There are different approaches to calculate reproduction rates (e. g.
\emph{Henneberg 1978}). We largely follow the methodology by
\emph{Hassan 1981}. Typically, a Total fertility rate (TFR) of 6-8 is
assumed for prehistoric populations (\emph{Ascadi/Nemeskeri 1970};
\emph{Henneberg 1978}; \emph{Hassan 1981}). Recently, \emph{McFadden
and Oxenham 2018} have published a formula to estimate the Total
fertility rate from archaeological data, provided that infants are
represented fully in the archaeological record.\cr
Unfortunately, this will not be the case for most archaeological
datasets. Therefore, we used the data published by McFadden and
Oxenham to apply it to the P(5-19)-index after \emph{Bocquet-Appel
2002}. We approximated the ratio by three different methods of
fitting (linear, logistic, power) and recommend logistic fitting,
but the others are available as well.\cr
The Gross reproduction rate (GRR) is calculated by multiplying the TFR
with the ratio of female newborns. The Net reproduction rate is
arrived at by summing the product of the GRR, the age specific
fertility rate as defined by \emph{Hassan} (1981, 137 tab. 8.7) and
the age specific survival taken from the life table and dividing the
result by 10000. The formulas for r and Dt are directly taken from
\emph{Hassan} (1981, 140).\cr
Also calculated is the ratio of dependent individuals which is usually
(but probably erroneously for archaic societies (\emph{Grupe et al.
2015}, 423) assumed to apply to those aged below 15 or 60 and above.
}
\examples{
# Calculate reproduction indices from real life dataset.
lt_aiterhofen <- life.table(aiterhofen_oedmuehlen[,2:3])
lt.reproduction(lt_aiterhofen)
}
\references{
\insertRef{acsadi_history_1970}{mortAAR}

\insertRef{bocquet_appel_2002}{mortAAR}

\insertRef{grupe_et_al_2015}{mortAAR}

\insertRef{hassan_1981}{mortAAR}

\insertRef{henneberg_1976}{mortAAR}

\insertRef{mcfadden_oxenham_2018}{mortAAR}
}
