% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{magdalenenberg}
\alias{magdalenenberg}
\title{magdalenenberg: Life table example}
\format{
A data frame with 13 rows and 2 variables.
\itemize{
  \item \bold{a:} age range.
  \item \bold{Dx:} number of deceased.
}
}
\description{
A data set from \emph{zaeuner_wahl_magdalenenberg_2013}.
}
\details{
Anthropological data from a large burial mound of the Early Iron Age in southwestern Germany. It is already
pooled for constructing a life table.
}
\references{
\insertRef{zaeuner_wahl_magdalenenberg_2013}{mortAAR}
}
\seealso{
Other lifetables: 
\code{\link{aiterhofen_oedmuehlen}},
\code{\link{gallery_graves}},
\code{\link{muensingen}},
\code{\link{nitra}},
\code{\link{odagsen_cm}},
\code{\link{odagsen_mo}},
\code{\link{schleswig_ma}}
}
\concept{lifetables}
