% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortAAR.R
\docType{package}
\name{mortAAR-package}
\alias{mortAAR}
\alias{mortAAR-package}
\title{mortAAR: Analysis of Archaeological Mortality Data}
\description{
A collection of functions for the analysis of archaeological mortality
data. It takes demographic data in different formats and displays the result in a standard life table
as well as plots the relevant indices.
}
\author{
\strong{Maintainer}: Nils Mueller-Scheessel \email{nils.mueller-scheessel@ufg.uni-kiel.de} [copyright holder]

Authors:
\itemize{
  \item Martin Hinz \email{martin.hinz@ufg.uni-kiel.de}
  \item Clemens Schmid \email{clemens@nevrome.de}
  \item Christoph Rinne \email{crinne@ufg.uni-kiel.de}
  \item Daniel Knitter \email{knitter@geographie.uni-kiel.de}
  \item Wolfgang Hamer \email{hamer@geographie.uni-kiel.de}
  \item Dirk Seidensticker \email{dirk.seidensticker@uni-tuebingen.de}
  \item Franziska Faupel \email{ffaupel@ufg.uni-kiel.de}
  \item Carolin Tietze \email{Carolin Tietze <ctietze1991@gmail.com>}
  \item Nicole Grunert \email{nicolegrunert@gmail.com}
}

}
\keyword{internal}
