% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{mortaar_plot_lx}
\alias{mortaar_plot_lx}
\title{Plots the survivorship lx for a single life table.}
\usage{
mortaar_plot_lx(x, lty = 1, ...)
}
\arguments{
\item{x}{an object of the class mortaar_life_table.}

\item{lty}{line type, default to 1.}

\item{...}{further arguments passed to the print function.}
}
\description{
Plots the survivorship lx for a single life table.
}
\keyword{internal}
