% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{mortaar_plot_lx_frame}
\alias{mortaar_plot_lx_frame}
\title{plots coordinate system for survivorship lx for a single life table}
\usage{
mortaar_plot_lx_frame(x, my_subsets = "", n, ...)
}
\arguments{
\item{x}{an object of the class mortaar_life_table}

\item{my_subsets}{a vector of categories from sublist of mortaar_life_table}

\item{n}{number of individuals}

\item{...}{further arguments passed to the print function}
}
\description{
plots coordinate system for survivorship lx for a single life table
}
