% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{mortaar_plot_qx}
\alias{mortaar_plot_qx}
\title{plots mortality rate qx for a single life table}
\usage{
mortaar_plot_qx(x, lty = 1, ...)
}
\arguments{
\item{x}{an object of the class mortaar_life_table}

\item{lty}{line type defaults to 1}

\item{...}{further arguments passed to the print function}
}
\description{
plots mortality rate qx for a single life table
}

