% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{mortaar_plot_x_axis}
\alias{mortaar_plot_x_axis}
\title{Calculates the x axis (age) for mortaar life table plots}
\usage{
mortaar_plot_x_axis(x)
}
\arguments{
\item{x}{an object of the class mortaar_life_table.}
}
\value{
the vector for the x axis
}
\description{
Calculates the x axis (age) for mortaar life table plots
}
\keyword{internal}
