% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nitra}
\alias{nitra}
\title{nitra: Life table example}
\format{
A data frame with 4 variables and 75 observations
\itemize{
  \item \bold{no:} individual
  \item \bold{sex:} sex determination if available
  \item \bold{age_start:} start of the ageing interval
  \item \bold{age_end:} end of the ageing interval (inclusive)
}
}
\description{
The Early Neolithic cemetery of Nitra is one of the best published
data sets of the \emph{Linearbandkeramik} (LBK). The individuals were
analysed by J. Jelínek (published in \emph{Pavúk 1972}) and many
decades later also bei L. Fibiger (\emph{Bickle/Whittle 2013}, 139ff.
Tab. 4.15) and in more detail by Z. \emph{Tvrdý 2016}. We follow here
the determinations by L. Fibiger.
}
\references{
\insertRef{pavuk_1972}{mortAAR}

\insertRef{LBK_lifeways}{mortAAR}

\insertRef{tvrdy_2016}{mortAAR}
}
\seealso{
Other lifetables: 
\code{\link{aiterhofen_oedmuehlen}},
\code{\link{gallery_graves}},
\code{\link{magdalenenberg}},
\code{\link{muensingen}},
\code{\link{odagsen_cm}},
\code{\link{odagsen_mo}},
\code{\link{schleswig_ma}}
}
\concept{lifetables}
