% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{plot.mortaar_life_table_list}
\alias{plot.mortaar_life_table_list}
\title{Plot a mortaar_life_table_list}
\usage{
\method{plot}{mortaar_life_table_list}(x, ...)
}
\arguments{
\item{x}{a mortaar_life_table_list}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot a mortaar_life_table_list
}

