% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{plot}
\alias{plot}
\alias{plot.mortaar_life_table}
\alias{plot.mortaar_life_table_list}
\title{Plot a mortaar_life_table or a mortaar_life_table_list}
\usage{
\method{plot}{mortaar_life_table}(x, display = c("dx", "qx", "lx", "ex",
  "rel_popx"), prefer.ggplot = TRUE, ...)

\method{plot}{mortaar_life_table_list}(x, display = c("dx", "qx", "lx", "ex",
  "rel_popx"), prefer.ggplot = TRUE, ...)
}
\arguments{
\item{x}{a mortaar_life_table or a mortaar_life_table_list.}

\item{display}{a character vector. Displays the calculated variables as plots. These must include
some of the alternatives \code{dx} for the proportion of deaths, \code{qx} for the probability of
death, \code{lx} for the survivorship, \code{ex} for the life expectancy and \code{rel_popx} for
the population age structure.}

\item{prefer.ggplot}{should ggplot be preferred, if available. Default to TRUE.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot a mortaar_life_table or a mortaar_life_table_list. There are several different kinds of plots
to explore the different variables. The parameter \code{display} allows to select the variables
of interest.
}
\examples{
limit = 100
steps = 5
lower <- seq(from = 0, to = limit-steps[1], by = steps)
upper <- seq(from = steps[1], to = limit, by = steps)-1

test <- list (
male = data.frame(
  x = paste0(lower, "--", upper),
  a = steps,
  Dx = runif(length(lower))*50
),
female = data.frame(
  x = paste0(lower, "--", upper),
  a = steps,
  Dx = runif(length(lower))*50
)
)
plot(life.table(test)$male)

}
