% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{print.mortaar_life_table_list}
\alias{print.mortaar_life_table_list}
\title{Print a mortaar_life_table_list}
\usage{
\method{print}{mortaar_life_table_list}(x, ...)
}
\arguments{
\item{x}{a mortaar_life_table_list}
}
\description{
Print a mortaar_life_table_list according to the format.mortaar_life_table_list
}
\examples{

test <- data.frame(male = round(runif(70)*100, 0), female = round(runif(70)*100, 0))
print.mortaar_life_table_list(life.table(test))

}

