% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_functions.R
\name{print}
\alias{print}
\alias{format.mortaar_life_table_list}
\alias{format.mortaar_life_table}
\alias{print.mortaar_life_table_list}
\alias{print.mortaar_life_table}
\title{Encode and print a mortaar_life_table or a mortaar_life_table_list}
\usage{
\method{format}{mortaar_life_table_list}(x, ...)

\method{format}{mortaar_life_table}(x, class_of_deceased = NULL, ...)

\method{print}{mortaar_life_table_list}(x, ...)

\method{print}{mortaar_life_table}(x, ...)
}
\arguments{
\item{x}{a mortaar_life_table or a mortaar_life_table_list.}

\item{...}{further arguments passed to or from other methods.}

\item{class_of_deceased}{optional string, specify the class of deceased (male, female, phase, ...).}
}
\value{
A string representation of the mortaar_life_table or the mortaar_life_table_list.
For format.mortaar_life_table_list each mortaar_life_table is formatted by itself and
strung together. The names of the elements are used to specify the name in the returned header
of the printout.
}
\description{
Format for pretty printing.
}
\examples{
# Create a mortaar_life_table from a prepared dataset.
schleswig_1 <- life.table(schleswig_ma[c("a", "Dx")])
print(schleswig_1)

# Create a mortaar_life_table_list from two datasets.
odagsen <- life.table(list(
  "corpus mandibulae" = odagsen_cm[c("a", "Dx")],
  "margo orbitalis" = odagsen_mo[c("a", "Dx")]
))
print(odagsen)

}
