% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_categories.R
\name{random.cat.apply}
\alias{random.cat.apply}
\title{Applying random age ranges to individuals with absolute ages}
\usage{
random.cat.apply(x, age, age_ranges, from, to)
}
\arguments{
\item{x}{a data.frame with individual absolute ages.}

\item{age_ranges}{a data.frame with age ranges.}

\item{from}{numeric. Column name for the begin of an age range.}

\item{to}{numeric. Column name for the end of an age range.}
}
\value{
The original data.frame \code{x} with two additional columns:

\itemize{
  \item \bold{from}:  Lower boundary of age category.
  \item \bold{to}: Upper boundary of age category.
 }
}
\description{
Helper function that applies random age categories to "known" absolute ages.
It is mainly used together with the functions \code{pop.sim.gomp}
and \code{random.cat}.
}
\examples{

# Simulate population and age ranges first
pop_sim <- pop.sim.gomp(n = 10000)
sim_ranges <- random.cat()

# apply random age categories to simulated ages
sim_appl <- random.cat.apply(pop_sim$result, age = "age", age_ranges = sim_ranges, from = "from", to = "to")
}
