% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_categories.R
\name{random.cat}
\alias{random.cat}
\title{Generation of random age ranges}
\usage{
random.cat(n_cat = 20, min_age = 15, max_cat_low = 60, max_age = 75)
}
\arguments{
\item{n_cat}{numeric. Number of categories, default: 20.}

\item{min_age}{numeric. Minimum age, default: 15.}

\item{max_cat_low}{numeric. Lower boundary of highest age categoriy, default: 60.}

\item{max_age}{numeric. Maximum age, default: 75.}
}
\value{
One data.frame with the following items:

\itemize{
  \item \bold{from}:  Lower boundary of age category.
  \item \bold{to}: Upper boundary of age category.
 }
}
\description{
Helper function that generates random age categories of absolute ages.
It is mainly used together with the functions \code{pop.sim.gomp}
and \code{random.cat.apply}.
}
\examples{
sim_ranges <- random.cat()
}
