% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{schleswig_ma}
\alias{schleswig_ma}
\title{schleswig_ma: Lifetable example}
\format{A data frame with 11 rows and 9 variables.
\itemize{
  \item x: fa, ma, sa -> fr체h-, mittel-, sp채tadult;
           fm, mm, sm -> fr체h-, mittel-, sp채tmatur;
           s -> senil
  \item a
  \item Dx
  \item dx
  \item lx
  \item qx
  \item Lx
  \item Tx
  \item ex
}
For further information about the values see the
documentation of \code{\link{life.table}}.}
\description{
A lifetable from \emph{Hermann et. al 1990, 305}.
}
\details{
The lifetable is compiled from a medieval grave collection from
Schleswig (11th-13th century). More information can possibly
be found in \emph{Grupe 1997}.
}
\references{
\insertRef{herrmann_prahistorische_1990}{mortAAR}

\insertRef{grupe_anthropologische_1997}{mortAAR}
}

