% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{summe}
\alias{summe}
\title{Uses the addition function "+" to combine two vectors setting NA values of the first vector to zero.}
\usage{
summe(x, y)
}
\arguments{
\item{x}{a vector.}

\item{y}{vector to combine with.}
}
\value{
dataframe with additional columns "from" and "to".
}
\description{
Uses the addition function "+" to combine two vectors setting NA values of the first vector to zero.
}
\examples{

#summentest=summe(c(1,2,NA,4),c(5,6,7,8)

}
\keyword{internal}
