% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{life.table}
\alias{life.table}
\title{Creates a life table}
\usage{
life.table(neclist)
}
\arguments{
\item{neclist}{list}
}
\value{
list
}
\description{
simple lifetable using Keyfitz and Flieger separation factors and
exponential tail of death distribution (to close out life table)
partly taken from https://web.stanford.edu/group/heeh/cgi-bin/web/node/75
}
\examples{
limit = 100
steps = 5
lower <- seq(from = 0, to = limit-steps[1], by = steps)
upper <- seq(from = steps[1], to = limit, by = steps)-1

neclist <- list (
 male = data.frame(
   x = paste0(lower, "--", upper),
   a = steps,
   Dx = runif(length(lower))*50
 ),
 female = data.frame(
   x = paste0(lower, "--", upper),
   a = steps,
   Dx = runif(length(lower))*50
 )
)

#neclist <- data.frame(male = round(runif(70)*100, 0), female = round(runif(70)*100, 0))

life.table(neclist)

}

