% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{life.table}
\alias{life.table}
\title{life table}
\usage{
life.table(neclist, agecorfac = c())
}
\arguments{
\item{neclist}{single dataframe or list of dataframes
             with the columns 'x', 'a', 'Dx'
\itemize{
  \item \bold{x}  age interval name, optional -
                  otherwise determined from \bold{a}
  \item \bold{a}  years within x
  \item \bold{Dx} number of deaths within x
}}

\item{agecorfac}{vector, optional. Age correction values to
determine the centre of the age interval for the
calculation of L(x). Given values replace the standard
values from the first age interval onwards.

Standard setup is: \eqn{agecorfac = a * \frac{1}{2}}.

Mainly used to correct higher mortality rates for
infants.}
}
\value{
An object of type mortaar_life_table or
mortaar_life_table_list. Each mortaar_life_table contains
the following variables:

\itemize{
  \item \bold{x}  age interval
  \item \bold{a}  years within x
  \item \bold{Dx} number of deaths within x
  \item \bold{dx} propotion of deaths within x (percent) :

                  \eqn{d_{x} = \frac{D_{x}}{\sum_{i=1}^{n} D_{i}} * 100}

  \item \bold{lx} survivorship within x (percent) :

                  \eqn{l_{x+1} = l_{x} - d_{x}} with \eqn{l_{0} = 100}

  \item \bold{qx} probability of death within x (percent) :

                  \eqn{q_{x} = \frac{d_{x}}{l_{x}}* 100}

  \item \bold{Ax} average number of years lived of an
                  individual that died within a specific
                  age class :

                  \eqn{A_{x} = a_{x} * agecorfac_{x}}

  \item \bold{Lx} average years per person lived within x :

                  \eqn{L_{x} = (a_{x} * l_{x}) - ((a_{x} - A_{x}) * d_{x})}

  \item \bold{Tx} sum of average years lived within
                  current and remaining x :

                  \eqn{T_{x+1} = T_{x} - L_{x}} with \eqn{T_{0} = \sum_{i=1}^{n}{L_{i}}}

  \item \bold{ex} average years of life remaining
                  (average life expectancy at mean(x)) :

                  \eqn{e_{x} = \frac{T_{x}}{l_{x}}}

  \item \bold{rel_popx} percentage of L(x) of the sum of L(x) :

                  \eqn{relpopx_{x} = \frac{L_{x}}{\sum_{i=1}^{n}{L_{i}}} * 100}
}
}
\description{
\code{life.table} calculates
\href{https://en.wikipedia.org/wiki/Life_table}{life table(s)}.
The algorithm is optimised for dead populations
encountered in archaeological research.
See \emph{Chamberlain 2006}, \emph{Hermann et. al 1990},
\emph{Kokkotidis/Richter 1991}, \emph{Keyfitz et al. 2005}
for the literature we examined. \cr
The function takes an individual data.frame or a list of
data.frames and respectively returns an object of type
mortaar_life_table or mortaar_life_table_list for which
exist specialised summary, print and plot functions.
}
\examples{

testdata1 <- schleswig_ma[c("a", "Dx")]

life.table(testdata1)
life.table(testdata1, c(0.25,1/3,0.5))

}
\references{
\insertRef{chamberlain_demography_2006}{mortAAR}

\insertRef{herrmann_prahistorische_1990}{mortAAR}

\insertRef{keyfitz_applied_2005}{mortAAR}

\insertRef{kokkotidis_graberfeld-_1991}{mortAAR}
}
