% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{life.table}
\alias{life.table}
\title{life table}
\usage{
life.table(neclist, acv = c())
}
\arguments{
\item{neclist}{list of dataframes or single dataframe
             with columns 'x', 'a', 'Dx'
\itemize{
  \item \bold{x}  age interval name (optional -
                  otherwise determined from \bold{a})
  \item \bold{a}  years within x
  \item \bold{Dx} number of deaths within x
}}

\item{acv}{vector, optional. Age correction values to
determine the centre of the age interval for the
calculation of L(x). Given values replace the standard
values from the first age interval onwards.

Standard setup is: \eqn{acv = a * \frac{1}{2}}.

Mainly used to correct higher mortality rates for
infants.}
}
\value{
Returns a list of dataframe(s), one for each life table
(males, females, sites etc.). Each dataframe with:

\itemize{
  \item \bold{x}  age interval
  \item \bold{a}  years within x
  \item \bold{Dx} number of deaths within x
  \item \bold{dx} propotion of deaths within x (percent) :

                  \eqn{d_{x} = \frac{D_{x}}{\sum_{i=1}^{n} D_{i}} * 100}

  \item \bold{lx} survivorship within x (percent) :

                  \eqn{l_{x+1} = l_{x} - d_{x}} with \eqn{l_{0} = 100}

  \item \bold{qx} probability of death within x (percent) :

                  \eqn{q_{x} = \frac{d_{x}}{l_{x}}* 100}

  \item \bold{Lx} average years per person lived within x :

                  \eqn{L_{x} = acv_{x} * (l_{x} + l_{x+1})}

  \item \bold{Tx} sum of average years lived within
                  current and remaining x :

                  \eqn{T_{x+1} = T_{x} - L_{x}} with \eqn{T_{0} = \sum_{i=1}^{n}{L_{i}}}

  \item \bold{ex} average years of life remaining
                  (average life expectancy at mean(x)) :

                  \eqn{e_{x} = \frac{T_{x}}{l_{x}}}

  \item \bold{Ax} percentage of L(x) of the sum of L(x) :

                  \eqn{A_{x} = \frac{L_{x}}{\sum_{i=1}^{n}{L_{i}}} * 100}
}
}
\description{
\code{life.table} returns a list of life table(s) for a
list of dataframe(s) with deseased (Dx) per age interval
(x)
}
\examples{

testdata1 <- schleswig_ma[c("a", "Dx")]

life.table(testdata1)
life.table(testdata1, c(0.25,1/3,0.5))

}
