% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{life.table}
\alias{life.table}
\title{Creates a life table}
\usage{
life.table(necdf)
}
\arguments{
\item{necdf}{data.frame}
}
\value{
list
}
\description{
simple lifetable using Keyfitz and Flieger separation factors and
exponential tail of death distribution (to close out life table)
partly taken from https://web.stanford.edu/group/heeh/cgi-bin/web/node/75
}
\examples{
test <- data.frame(male = round(runif(70)*100, 0), female = round(runif(70)*100, 0))

life.table(test)

}

