% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetable_indices.R
\name{lt.mortality}
\alias{lt.mortality}
\alias{lt.mortality.default}
\alias{lt.mortality.mortaar_life_table_list}
\alias{lt.mortality.mortaar_life_table}
\title{Generates mortality indices from a mortAAR life table for use in other functions}
\usage{
lt.mortality(life_table)

\method{lt.mortality}{default}(life_table)

\method{lt.mortality}{mortaar_life_table_list}(life_table)

\method{lt.mortality}{mortaar_life_table}(life_table)
}
\arguments{
\item{life_table}{an object of class mortaar_life_table.}
}
\value{
A list with the following:
\itemize{
  \item \bold{mortality}:   list of mortalities 5q0, 5q10, 5q15, 45q15.
}
}
\description{
This function computes mortality indices from a mortAAR
life table which are needed by other functions in mortAAR.
}
