% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortAAR.R
\docType{package}
\name{mortAAR-package}
\alias{mortAAR}
\alias{mortAAR-package}
\title{mortAAR: Analysis of Archaeological Mortality Data}
\description{
A collection of functions for the analysis of archaeological mortality data (on the topic see e.g. Chamberlain 2006 \url{https://books.google.de/books?id=nG5FoO_becAC&lpg=PA27&ots=LG0b_xrx6O&dq=life\%20table\%20archaeology&pg=PA27#v=onepage&q&f=false}). It takes demographic data in different formats and displays the result in a standard life table as well as plots the relevant indices (percentage of deaths, survivorship, probability of death, life expectancy, percentage of population).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ISAAKiel/mortAAR}
  \item Report bugs at \url{https://github.com/ISAAKiel/mortAAR/issues}
}

}
\author{
\strong{Maintainer}: Nils Mueller-Scheessel \email{nils.mueller-scheessel@ufg.uni-kiel.de} [copyright holder]

Authors:
\itemize{
  \item Martin Hinz \email{martin.hinz@ufg.uni-kiel.de}
  \item Clemens Schmid \email{clemens@nevrome.de}
  \item Christoph Rinne \email{crinne@ufg.uni-kiel.de}
  \item Daniel Knitter \email{knitter@geographie.uni-kiel.de}
  \item Wolfgang Hamer \email{hamer@geographie.uni-kiel.de}
  \item Dirk Seidensticker \email{dirk.seidensticker@uni-tuebingen.de}
  \item Franziska Faupel \email{ffaupel@ufg.uni-kiel.de}
  \item Carolin Tietze \email{ctietze1991@gmail.com}
  \item Nicole Grunert \email{nicolegrunert@gmail.com}
}

}
\keyword{internal}
