% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{mortaar_plot_qx}
\alias{mortaar_plot_qx}
\title{Plots the probability of death qx for a single life table.}
\usage{
mortaar_plot_qx(x, lty = 1, ...)
}
\arguments{
\item{x}{an object of the class mortaar_life_table.}

\item{lty}{line type, default to 1.}

\item{...}{further arguments passed to the print function.}
}
\description{
Plots the probability of death qx for a single life table.
}
\keyword{internal}
