% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{plot.mortaar_life_table}
\alias{plot.mortaar_life_table}
\title{Plot a mortaar_life_table}
\usage{
\method{plot}{mortaar_life_table}(x, ...)
}
\arguments{
\item{x}{a mortaar_life_table}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot a mortaar_life_table according to the format.mortaar_life_table
}
\examples{

test <- data.frame(male = round(runif(70)*100, 0), female = round(runif(70)*100, 0))
plot(life.table(test)$male)

}

