% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{plot.mortaar_life_table_list}
\alias{plot.mortaar_life_table_list}
\title{Plot a mortaar_life_table_list}
\usage{
\method{plot}{mortaar_life_table_list}(x, display = c("qx", "ex", "Ax"), ...)
}
\arguments{
\item{x}{a mortaar_life_table_list}

\item{display}{which plots to show. These must include some of the
alternatives \code{qx} for survivorship, \code{ex} for mortality rate
and \code{Ax} for population age structure.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot a mortaar_life_table_list
}
\examples{
limit = 100
steps = 5
lower <- seq(from = 0, to = limit-steps[1], by = steps)
upper <- seq(from = steps[1], to = limit, by = steps)-1

test <- list (
male = data.frame(
  x = paste0(lower, "--", upper),
  a = steps,
  Dx = runif(length(lower))*50
),
female = data.frame(
  x = paste0(lower, "--", upper),
  a = steps,
  Dx = runif(length(lower))*50
)
)
plot(life.table(test))

}
