% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{prep.life.table}
\alias{prep.life.table}
\title{Creates the input for the function life.table}
\usage{
prep.life.table(x, dec = "NA", agebeg, ageend, grnam = "NA",
  methode = "NA")
}
\arguments{
\item{x}{a dataframe}

\item{dec}{Column name of the count of deceased (as character)}

\item{agebeg}{Column name of the from field (as character)}

\item{ageend}{Column name of the to field (as character)}

\item{grnam}{Column name of the grouping field (as character)}

\item{methode}{name of the age class determination (as character) Options: "Standard" (default) (1,4,5,5,...), "Equal5" (5,5,...)}
}
\value{
list as input parameter for the function life.table
}
\description{
Creates the input for the function life.table
}
\examples{

#test2=prep.life.table(x=test1,dec="NA",agebeg="from",
#ageend="to",grnam="Geschlecht", methode="Standard")

}
