% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{schleswig_ma}
\alias{schleswig_ma}
\title{schleswig_ma: Life table example}
\format{A data frame with 11 rows and 9 variables.
\itemize{
  \item \bold{x:} fa, ma, sa -> fr체h-, mittel-, sp채tadult;
                  fm, mm, sm -> fr체h-, mittel-, sp채tmatur;
                  s -> senil
  \item \bold{a}
  \item \bold{Dx}
  \item \bold{dx}
  \item \bold{lx}
  \item \bold{qx}
  \item \bold{Lx}
  \item \bold{Tx}
  \item \bold{ex}
}
For further information about the values see the
documentation of \code{\link{life.table}}.}
\description{
A life table from \emph{Herrmann et. al 1990, 305}.
}
\details{
This lifetable is compiled from a medieval grave collection from
Schleswig (11th-13th century). More information can
be found in \emph{Grupe 1997}.
}
\references{
\insertRef{herrmann_prahistorische_1990}{mortAAR}

\insertRef{grupe_anthropologische_1997}{mortAAR}
}
\seealso{
Other lifetables: \code{\link{aiterhofen_oedmuehlen}},
  \code{\link{odagsen_cm}}, \code{\link{odagsen_mo}}
}
\concept{lifetables}
